+++ tdrive/backend/node/src/core/platform/services/database/services/orm/connectors/postgres/postgres.ts
@@ -50,6 +50,12 @@ export class PostgresConnector extends AbstractConnector<PostgresConnectionOptio
     return this;
   }
 
+  async disconnect(): Promise<this> {
+    if (this.client) await this.client.end();
+    this.client = null;
+    return this;
+  }
+
   async init(): Promise<this> {
     if (!this.client) {
       await this.connect();
@@ -165,7 +171,18 @@ export class PostgresConnector extends AbstractConnector<PostgresConnectionOptio
   }
 
   async drop(): Promise<this> {
-    logger.info("Drop database is not implemented for security reasons.");
+    const query = `
+        DO $$ 
+        DECLARE 
+          tablename text;
+        BEGIN 
+          FOR tablename IN (SELECT table_name FROM information_schema.tables WHERE table_schema = 'public') 
+          LOOP 
+            EXECUTE 'DELETE FROM  "' || tablename || '" CASCADE'; 
+          END LOOP; 
+        END $$;`;
+    logger.debug(`service.database.orm.postgres.drop - Query: "${query}"`);
+    await this.client.query(query);
     return this;
   }