+++ tdrive/backend/node/src/core/platform/services/search/adapters/elasticsearch/elastic-open-search-adapter.ts
@@ -29,6 +29,7 @@ export default class ESAndOpenSearch extends SearchAdapter implements SearchAdap
   private bulkReaders = 0;
   private buffer: Operation[] = [];
   private client: OpenClient | ESClient;
+  private pendingTimeout: NodeJS.Timeout = null;
 
   constructor(
     readonly database: DatabaseServiceAPI,
@@ -73,6 +74,18 @@ export default class ESAndOpenSearch extends SearchAdapter implements SearchAdap
       );
     }
     this.startBulkReader();
+  }
+
+  public async disconnect() {
+    if (this.client) {
+      const client = this.client;
+      this.client = null;
+      if (this.pendingTimeout) {
+        clearTimeout(this.pendingTimeout);
+        this.pendingTimeout = null;
+      }
+      await client.close();
+    }
   }
 
   private async createIndex(
@@ -214,10 +227,17 @@ export default class ESAndOpenSearch extends SearchAdapter implements SearchAdap
 
     let buffer;
     do {
-      await new Promise(r =>
-        setTimeout(r, parseInt(`${this.configuration.flushInterval}`) || 3000),
+      await new Promise<void>(
+        r =>
+          (this.pendingTimeout = setTimeout(() => {
+            this.pendingTimeout = null;
+            r();
+          }, parseInt(`${this.configuration.flushInterval}`) || 3000)),
       );
       buffer = this.buffer;
+      if (!this.client)
+        // disconnect was called; break this tail loop
+        return;
     } while (buffer.length === 0);
     this.buffer = [];
 
@@ -283,11 +303,12 @@ export default class ESAndOpenSearch extends SearchAdapter implements SearchAdap
     const esParamsWithScroll = {
       ...esParams,
       size: parseInt(options.pagination.limitStr || "100"),
-      scroll: "1m",
+      scroll: "10m",
     };
 
     let esResponse: any;
 
+    let nextToken;
     if (options.pagination.page_token) {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
@@ -297,18 +318,23 @@ export default class ESAndOpenSearch extends SearchAdapter implements SearchAdap
         },
         esOptions,
       );
+      //the scroll token is also the same, and we do not get it from response
+      nextToken = options.pagination.page_token;
     } else {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       esResponse = await this.client.search(esParamsWithScroll, esOptions);
+      nextToken = esResponse.body?._scroll_id || "";
     }
 
     if (esResponse.statusCode !== 200) {
       logger.error(`${this.name} -  ${JSON.stringify(esResponse.body)}`);
     }
 
-    const nextToken = esResponse.body?._scroll_id || "";
     const hits = esResponse.body?.hits?.hits || [];
+    if (hits.length == 0) {
+      nextToken = false;
+    }
 
     logger.debug(`${this.name} got response: ${JSON.stringify(esResponse)}`);