+++ tdrive/backend/node/src/services/documents/services/index.ts
@@ -2,8 +2,8 @@ import SearchRepository from "../../../core/platform/services/search/repository"
 import { getLogger, logger, TdriveLogger } from "../../../core/platform/framework";
 import {
   CrudException,
-  ExecutionContext,
   ListResult,
+  Pagination,
 } from "../../../core/platform/framework/api/crud-service";
 import Repository, {
   comparisonType,
@@ -60,7 +60,7 @@ import {
   RealtimeEntityActionType,
   ResourcePath,
 } from "../../../core/platform/services/realtime/types";
-
+import config from "config";
 export class DocumentsService {
   version: "1";
   repository: Repository<DriveFile>;
@@ -69,6 +69,12 @@ export class DocumentsService {
   driveTdriveTabRepository: Repository<DriveTdriveTabEntity>;
   ROOT: RootType = "root";
   TRASH: TrashType = "trash";
+  quotaEnabled: boolean = config.has("drive.featureUserQuota")
+    ? config.get("drive.featureUserQuota")
+    : false;
+  defaultQuota: number = config.has("drive.defaultUserQuota")
+    ? config.get("drive.defaultUserQuota")
+    : 0;
   logger: TdriveLogger = getLogger("Documents Service");
 
   async init(): Promise<this> {
@@ -100,19 +106,32 @@ export class DocumentsService {
     context: DriveExecutionContext & { public_token?: string },
   ): Promise<BrowseDetails> => {
     if (isSharedWithMeFolder(id)) {
-      const children = await this.search(options, context);
-      return {
-        access: "read",
-        children: children.getEntities(),
-        nextPage: children.nextPage,
-        path: [] as Array<DriveFile>,
-      };
+      return this.sharedWithMe(options, context);
     } else {
       return {
         nextPage: null,
         ...(await this.get(id, context)),
       };
     }
+  };
+
+  sharedWithMe = async (
+    options: SearchDocumentsOptions,
+    context: DriveExecutionContext & { public_token?: string },
+  ): Promise<BrowseDetails> => {
+    const result = [];
+    let fileList: ListResult<DriveFile>;
+    do {
+      fileList = await this.search(options, context);
+      result.push(...fileList.getEntities());
+      options.pagination = fileList.nextPage;
+    } while (fileList.nextPage?.page_token);
+    return {
+      access: "read",
+      children: result,
+      nextPage: null,
+      path: [] as Array<DriveFile>,
+    };
   };
 
   userQuota = async (context: CompanyExecutionContext): Promise<number> => {
@@ -315,6 +334,13 @@ export class DocumentsService {
         }
 
         if (fileToProcess) {
+          // if quota is enabled, check if the user has enough space
+          await this.checkQuotaAndCleanUp(
+            fileToProcess.upload_data.size,
+            fileToProcess.id,
+            context,
+          );
+
           driveItem.size = fileToProcess.upload_data.size;
           driveItem.is_directory = false;
           driveItem.extension = fileToProcess.metadata.name.split(".").pop();
@@ -356,19 +382,17 @@ export class DocumentsService {
               )
             : [];
 
-          if (sharedWith.length > 0) {
-            // Notify the user that the document has been shared with them
-            this.logger.info("Notifying users that the document has been shared with them: ", {
+          if (context.user.id !== parentItem?.creator && sharedWith.length > 0) {
+            // Notify the owner that the document has been shared with them
+            this.logger.info("Notifying the onwer that the document has been shared with them: ", {
               sharedWith,
             });
-            for (const info of sharedWith) {
-              gr.services.documents.engine.notifyDocumentShared({
-                context,
-                item: driveItem,
-                notificationEmitter: context.user.id,
-                notificationReceiver: info.id,
-              });
-            }
+            gr.services.documents.engine.notifyDocumentShared({
+              context,
+              item: driveItem,
+              notificationEmitter: context.user.id,
+              notificationReceiver: parentItem.creator,
+            });
           }
         }
       } catch (error) {
@@ -456,6 +480,7 @@ export class DocumentsService {
         if ((content as any)[key]) {
           if (
             key === "parent_id" &&
+            oldParent !== item.parent_id &&
             !(await canMoveItem(item.id, content.parent_id, this.repository, context))
           ) {
             throw Error("Move operation not permitted");
@@ -465,8 +490,9 @@ export class DocumentsService {
           if (key === "access_info") {
             const sharedWith = content.access_info.entities.filter(
               info =>
-                !item.access_info.entities.find(entity => entity.id === info.id) &&
-                info.type === "user",
+                info.type === "user" &&
+                info.id !== context.user.id &&
+                !item.access_info.entities.find(entity => entity.id === info.id),
             );
 
             item.access_info = content.access_info;
@@ -484,7 +510,7 @@ export class DocumentsService {
               });
             }
 
-            item.access_info.entities.forEach(async info => {
+            item.access_info.entities.forEach(info => {
               if (!info.grantor) {
                 info.grantor = context.user.id;
               }
@@ -517,13 +543,10 @@ export class DocumentsService {
 
       if (oldParent) {
         item.scope = await getItemScope(item, this.repository, context);
-        this.repository.save(item);
+        await this.repository.save(item);
 
         await updateItemSize(oldParent, this.repository, context);
-        this.notifyWebsocket(oldParent, context);
       }
-
-      this.notifyWebsocket(item.parent_id, context);
 
       if (item.parent_id === this.TRASH) {
         //When moving to trash we recompute the access level to make them flat
@@ -669,6 +692,7 @@ export class DocumentsService {
    * restore a Drive Document and its children
    *
    * @param {string} id - the item id
+   * @param item item to restore
    * @param {DriveExecutionContext} context - the execution context
    * @returns {Promise<void>}
    */
@@ -704,7 +728,7 @@ export class DocumentsService {
       throw new CrudException("User does not have access to this item or its children", 401);
     }
 
-    if (isInTrash(item, this.repository, context)) {
+    if (await isInTrash(item, this.repository, context)) {
       if (item.is_in_trash != true) {
         if (item.scope === "personal") {
           item.parent_id = "user_" + context.user.id;
@@ -715,9 +739,7 @@ export class DocumentsService {
         item.is_in_trash = false;
       }
     }
-    this.repository.save(item);
-
-    this.notifyWebsocket("trash", context);
+    await this.repository.save(item);
   };
 
   /**
@@ -765,6 +787,9 @@ export class DocumentsService {
       const driveItemVersion = getDefaultDriveItemVersion(version, context);
       const metadata = await getFileMetadata(driveItemVersion.file_metadata.external_id, context);
 
+      // if quota is enabled, check if the user has enough space
+      await this.checkQuotaAndCleanUp(metadata.size, metadata.external_id, context);
+
       driveItemVersion.file_size = metadata.size;
       driveItemVersion.file_metadata.size = metadata.size;
       driveItemVersion.file_metadata.name = metadata.name;
@@ -783,21 +808,22 @@ export class DocumentsService {
       await this.repository.save(item);
 
       // Notify the user that the document versions have been updated
-      this.logger.info("Notifying user that the document has been updated: ", {
-        item,
-        notificationEmitter: context.user.id,
-      });
-      gr.services.documents.engine.notifyDocumentVersionUpdated({
-        context,
-        item,
-        notificationEmitter: context.user.id,
-        notificationReceiver: item.creator,
-      });
+      if (context.user.id !== item.creator) {
+        this.logger.info("Notifying user that the document has been updated: ", {
+          item,
+          notificationEmitter: context.user.id,
+        });
+        gr.services.documents.engine.notifyDocumentVersionUpdated({
+          context,
+          item,
+          notificationEmitter: context.user.id,
+          notificationReceiver: item.creator,
+        });
+      }
 
-      this.notifyWebsocket(item.parent_id, context);
       await updateItemSize(item.parent_id, this.repository, context);
 
-      globalResolver.platformServices.messageQueue.publish<DocumentsMessageQueueRequest>(
+      await globalResolver.platformServices.messageQueue.publish<DocumentsMessageQueueRequest>(
         "services:documents:process",
         {
           data: {
@@ -810,8 +836,8 @@ export class DocumentsService {
 
       return driveItemVersion;
     } catch (error) {
-      this.logger.error({ error: `${error}` }, "Failed to create Drive item version");
-      throw new CrudException("Failed to create Drive item version", 500);
+      logger.error({ error: `${error}` }, "Failed to create Drive item version");
+      CrudException.throwMe(error, new CrudException("Failed to create Drive item version", 500));
     }
   };
 
@@ -913,6 +939,10 @@ export class DocumentsService {
       zlib: { level: 9 },
     });
 
+    archive.on("error", error => {
+      this.logger.error("error while creating ZIP file: ", error);
+    });
+
     for (const id of ids) {
       if (!(await checkAccess(id, null, "read", this.repository, context))) {
         this.logger.warn(`not enough permissions to download ${id}, skipping`);
@@ -927,6 +957,7 @@ export class DocumentsService {
       }
     }
 
+    //TODO[ASH] why do we need this call??
     archive.finalize();
 
     return archive;
@@ -959,9 +990,9 @@ export class DocumentsService {
     const result = await this.searchRepository.search(
       {},
       {
-        pagination: {
-          limitStr: "100",
-        },
+        pagination: options.pagination
+          ? Pagination.fromPaginable(options.pagination)
+          : new Pagination(),
         $in: [
           ...(options.onlyDirectlyShared ? [["access_entities", [context.user.id]] as inType] : []),
           ...(options.company_id ? [["company_id", [options.company_id]] as inType] : []),
@@ -975,6 +1006,7 @@ export class DocumentsService {
               ]
             : []),
         ],
+        $nin: [...(options.onlyUploadedNotByMe ? [["creator", [context.user.id]] as inType] : [])],
         $lte: [
           ...(options.last_modified_lt
             ? [["last_modified", options.last_modified_lt] as comparisonType]
@@ -1003,8 +1035,11 @@ export class DocumentsService {
     if (!options.onlyDirectlyShared) {
       const filteredResult = await this.filter(result.getEntities(), async item => {
         try {
-          // Check access for each item
-          return await checkAccess(item.id, null, "read", this.repository, context);
+          //skip all the fiels
+          return (
+            !item.is_in_trash &&
+            (await checkAccess(item.id, null, "read", this.repository, context))
+          );
         } catch (error) {
           this.logger.warn("failed to check item access", error);
           return false;
@@ -1016,12 +1051,11 @@ export class DocumentsService {
 
     if (options.onlyUploadedNotByMe) {
       const filteredResult = await this.filter(result.getEntities(), async item => {
-        return item.creator != context.user.id;
+        return item.creator != context.user.id && !item.is_in_trash;
       });
 
       return new ListResult(result.type, filteredResult, result.nextPage);
     }
-
     return result;
   };
 
@@ -1099,5 +1133,17 @@ export class DocumentsService {
     );
 
     return await this.getTab(tabId, context);
+  };
+
+  checkQuotaAndCleanUp = async (size: number, fileId: string, context: DriveExecutionContext) => {
+    if (!this.quotaEnabled) return;
+
+    const userQuota = await this.userQuota(context);
+    const leftQuota = this.defaultQuota - userQuota;
+
+    if (size > leftQuota) {
+      await globalResolver.services.files.delete(fileId, context);
+      throw new CrudException(`Not enough space: ${size}, ${leftQuota}.`, 403);
+    }
   };
 }