+++ tdrive/backend/node/src/services/documents/web/controllers/documents.ts
@@ -6,9 +6,9 @@ import { File } from "../../../../services/files/entities/file";
 import { UploadOptions } from "../../../../services/files/types";
 import globalResolver from "../../../../services/global-resolver";
 import {
+  CompanyUserRole,
   PaginationQueryParameters,
   ResourceWebsocket,
-  CompanyUserRole,
 } from "../../../../utils/types";
 import { DriveFile } from "../../entities/drive-file";
 import { FileVersion } from "../../entities/file-version";
@@ -50,7 +50,7 @@ export class DocumentsController {
         version: Partial<FileVersion>;
       };
     }>,
-  ): Promise<DriveFile> => {
+  ): Promise<DriveFile | any> => {
     try {
       const context = getDriveExecutionContext(request);
 
@@ -73,6 +73,7 @@ export class DocumentsController {
 
       const { item, version } = request.body;
 
+      //
       return await globalResolver.services.documents.documents.create(
         createdFile,
         item,
@@ -199,48 +200,10 @@ export class DocumentsController {
       onlyUploadedNotByMe: true,
     };
 
-    const data = {
+    return {
       ...(await globalResolver.services.documents.documents.browse(id, options, context)),
-      websockets: request.currentUser?.id
-        ? globalResolver.platformServices.realtime.sign(
-            [{ room: `/companies/${context.company.id}/documents/item/${id}` }],
-            request.currentUser?.id,
-          )
-        : [],
+      websockets: [],
     };
-
-    return data;
-  };
-
-  sharedWithMe = async (
-    request: FastifyRequest<{
-      Params: RequestParams;
-      Body: SearchDocumentsBody;
-      Querystring: { public_token?: string };
-    }>,
-  ): Promise<ListResult<DriveFileDTO>> => {
-    try {
-      const context = getDriveExecutionContext(request);
-
-      const options: SearchDocumentsOptions = {
-        ...request.body,
-        company_id: request.body.company_id || context.company.id,
-        view: DriveFileDTOBuilder.VIEW_SHARED_WITH_ME,
-        onlyDirectlyShared: true,
-        onlyUploadedNotByMe: true,
-      };
-
-      if (!Object.keys(options).length) {
-        this.throw500Search();
-      }
-
-      const fileList = await globalResolver.services.documents.documents.search(options, context);
-
-      return this.driveFileDTOBuilder.build(fileList, context, options.fields, options.view);
-    } catch (error) {
-      logger.error({ error: `${error}` }, "error while searching for document");
-      this.throw500Search();
-    }
   };
 
   /**
@@ -343,14 +306,19 @@ export class DocumentsController {
       Body: Partial<FileVersion>;
       Querystring: { public_token?: string };
     }>,
-  ): Promise<FileVersion> => {
-    const context = getDriveExecutionContext(request);
-    const { id } = request.params;
-    const version = request.body;
+  ): Promise<FileVersion | any> => {
+    try {
+      const context = getDriveExecutionContext(request);
+      const { id } = request.params;
+      const version = request.body;
 
-    if (!id) throw new CrudException("Missing id", 400);
+      if (!id) throw new CrudException("Missing id", 400);
 
-    return await globalResolver.services.documents.documents.createVersion(id, version, context);
+      return await globalResolver.services.documents.documents.createVersion(id, version, context);
+    } catch (error) {
+      logger.error({ error: `${error}` }, "Failed to create Drive item version");
+      CrudException.throwMe(error, new CrudException("Failed to create Drive item version", 500));
+    }
   };
 
   downloadGetToken = async (