+++ tdrive/frontend/src/app/features/users/hooks/use-user-quota.ts
@@ -3,6 +3,8 @@ import { UserQuota } from '@features/users/types/user-quota';
 import UserAPIClient from '@features/users/api/user-api-client';
 import { useCurrentUser } from "features/users/hooks/use-current-user";
 import { atom, useRecoilState } from "recoil";
+import useRouteState from "app/features/router/hooks/use-route-state";
+import useRouterCompany from "app/features/router/hooks/use-router-company";
 
 export const QuotaState = atom<UserQuota>({
   key: 'QuotaState',
@@ -19,13 +21,16 @@ export const useUserQuota = () => {
     remaining: 1,
     total: 1
   }
-  const {user } = useCurrentUser();
+  const { appName } = useRouteState();
+  const isPublic = appName === 'drive';
+  const companyId = useRouterCompany();
+  const { user } = isPublic ? { user: null } : useCurrentUser();
   const [quota, setQuota] = useRecoilState(QuotaState);
 
   const getQuota = useCallback(async () => {
     let data: UserQuota = nullQuota;
-    if (user?.id) {
-      data = await UserAPIClient.getQuota(user.id);
+    if (user && user?.id) {
+      data = await UserAPIClient.getQuota(companyId, user.id);
     } else {
       data = nullQuota;
     }