+++ tdrive/frontend/src/app/views/client/body/drive/header-path.tsx
@@ -15,10 +15,12 @@ export default ({
   path: livePath,
   inTrash,
   setParentId,
+  inPublicSharing,
 }: {
   path: DriveItem[];
   inTrash?: boolean;
   setParentId: (id: string) => void;
+  inPublicSharing?: boolean;
 }) => {
   const [savedPath, setSavedPath] = useState<DriveItem[]>([]);
   const history = useHistory();
@@ -40,7 +42,7 @@ export default ({
             dirId,
           }),
         );
-        setParentId(dirId ? dirId : viewId);
+        if (inPublicSharing) return setParentId(dirId ? dirId : viewId);
       }}
     />
   );
@@ -128,7 +130,7 @@ const PathItem = ({
   const { user } = useCurrentUser();
   const { viewId } = RouterServices.getStateFromRoute();
   const { access: trashAccess } = useDriveItem('trash');
-  const isInSharedWithMe = viewId === "shared_with_me";
+  const isInSharedWithMe = viewId === 'shared_with_me';
   return (
     <div className="flex items-center">
       <a
@@ -136,18 +138,6 @@ const PathItem = ({
         className="text-sm font-medium text-gray-700 hover:text-blue-600 dark:text-gray-400 dark:hover:text-white"
         onClick={evt => {
           evt.preventDefault();
-          const driveMenuItems = [
-            {
-              type: 'menu',
-              text: Languages.t('components.side_menu.home'),
-              onClick: () => onClick('root', ''),
-            },
-            {
-              type: 'menu',
-              text: Languages.t('components.side_menu.my_drive'),
-              onClick: () => onClick('user_' + user?.id, ''),
-            },
-          ];
 
           const trashMenuItems = [
             {
@@ -171,7 +161,11 @@ const PathItem = ({
             if (viewId?.includes('trash')) {
               MenusManager.openMenu(trashMenuItems, { x: evt.clientX, y: evt.clientY }, 'center');
             } else {
-              MenusManager.openMenu(driveMenuItems, { x: evt.clientX, y: evt.clientY }, 'center');
+              if (viewId === 'root') {
+                onClick('root', '')
+              } else if (viewId === 'user_' + user?.id) {
+                onClick('user_' + user?.id, '')
+              }
             }
           } else {
             onClick(viewId || '', item?.id || '');
@@ -185,12 +179,12 @@ const PathItem = ({
 
             if (first) {
               if (isTrash) {
-                return viewId?.includes('trash_') 
-                  ? Languages.t('components.header_path.my_trash') 
+                return viewId?.includes('trash_')
+                  ? Languages.t('components.header_path.my_trash')
                   : Languages.t('components.header_path.shared_trash');
               } else {
-                return isInSharedWithMe 
-                  ? Languages.t('components.header_path.shared_with_me') 
+                return isInSharedWithMe
+                  ? Languages.t('components.header_path.shared_with_me')
                   : fileName;
               }
             } else {
@@ -202,7 +196,7 @@ const PathItem = ({
       {item?.access_info?.public?.level && item?.access_info?.public?.level !== 'none' && (
         <PublicIcon className="h-5 w-5 ml-2" />
       )}
-      {first && !!user?.id && viewId != "shared_with_me" && (
+      {first && !!user?.id && viewId?.includes('trash') && (
         <span className="ml-2 -mr-1 text-gray-700">
           <ChevronDownIcon className="w-4 h-4" />
         </span>